api/requirements.txt
cryptography==41.0.1
fastapi==0.96.0
pydantic==1.10.8
PyJWT==2.7.0
PyJWT==2.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.15
starlette<0.28.0,>=0.27.0
uvicorn==0.15.0